feat: add dns-rebinding & http log (#3)
* h-w

* feat: add dns-rebinding

* fixs

* fix

* docs

+++ api/util.go
@@ -6,7 +6,7 @@ import (
 
 // RespJSON respsone json struct
 type RespJSON struct {
-	Code    uint16      `json:"code"`
+	Code    int         `json:"code"`
 	Message string      `json:"message"`
 	Data    interface{} `json:"data"`
 	Success bool        `json:"success"`
